rearranged directory structure
