+++ lib/views/full_setup/full_setup.dart
@@ -1,0 +1,1 @@
+export 'views/subjects.dart';
