+++ lib/views/full_setup/views/subjects.dart
@@ -2,9 +2,9 @@ import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_platform_widgets/flutter_platform_widgets.dart';
 
-import '../../models/models.dart';
-import '../../services/services.dart';
-import '../../styles/styles.dart';
+import '../../../models/models.dart';
+import '../../../services/services.dart';
+import '../../../styles/styles.dart';
 
 class Subjects extends StatelessWidget {
   final UserModel user;
