+++ lib/views/initial_setup/initial_setup.dart
@@ -1,0 +1,2 @@
+export 'views/universities.dart';
+export 'views/welcome.dart';
