+++ lib/views/initial_setup/views/local_widgets/local_widgets.dart
@@ -1,0 +1,1 @@
+export 'university_list.dart';
