+++ lib/views/initial_setup/views/local_widgets/university_list.dart
@@ -4,31 +4,11 @@ import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_platform_widgets/flutter_platform_widgets.dart';
 
-import '../../models/models.dart';
-import '../../services/services.dart';
-import '../../shared/shared.dart';
-import '../../styles/styles.dart';
-import '../views.dart';
-
-class Universities extends StatelessWidget {
-  @override
-  Widget build(BuildContext context) {
-    return FutureBuilder<List<University>>(
-      future:
-          Collection<University>(path: 'hochschulen', order: ['name', 'ASC'])
-              .getData(),
-      builder:
-          (BuildContext context, AsyncSnapshot<List<University>> snapshot) {
-        if (snapshot.connectionState == ConnectionState.done) {
-          if (snapshot.hasData) {
-            return UniversityList(universities: snapshot.data);
-          }
-        }
-        return const LoadingScreen();
-      },
-    );
-  }
-}
+import '../../../../models/models.dart';
+import '../../../../services/services.dart';
+import '../../../../shared/shared.dart';
+import '../../../../styles/styles.dart';
+import '../../../views.dart';
 
 class UniversityList extends StatefulWidget {
   final List<University> universities;
