+++ lib/views/views.dart
@@ -1,8 +1,8 @@
 export 'cafeteria/cafeteria.dart';
 export 'calendar/calendar.dart';
-export 'full_setup/subjects.dart';
+export 'full_setup/full_setup.dart';
 export 'home.dart';
-export 'initial_setup/welcome.dart';
+export 'initial_setup/initial_setup.dart';
 export 'news/news.dart';
 export 'organisation/organisation.dart';
 export 'settings/settings.dart';
